pyproject.toml
[tool.poetry]
name = "quo-client"
version = "2.0.0"
description = "A client library for accessing Quo (OpenPhone) Public API"
authors = []
readme = "README.md"
packages = [
    {include = "quo_client"},
    { include = "quo_sdk" },
]
include = ["CHANGELOG.md", "quo_client/py.typed"]


# ─── dependencies ──────────────────────────────────────────────
[tool.poetry.dependencies]
python = "^3.9"
httpx  = ">=0.23.0,<0.29.0"
attrs  = ">=22.2.0"
python-dateutil = "^2.8.0"

# ─── dev-only deps (older Poetry uses the dash, not a dot) ────
python-dotenv = "^1.1.0"
pytz = "^2025.2"
[tool.poetry.dev-dependencies]
pytest       = "^8.2"
pytest-httpx = "^0.30"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]

[tool.mypy]
strict = true

